requirements.txt
kaggle
torch
torchvision
scikit-image
pillow
Flask
numpy
matplotlib
pycocotools
opencv-python==4.9.0.80